% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializecisTopic.R
\name{addCellMetadata}
\alias{addCellMetadata}
\title{Add cells metadata to a cisTopic object}
\usage{
addCellMetadata(object, cell.data, ...)
}
\arguments{
\item{object}{cisTopic object}

\item{cell.data}{Additional metadata to add to the cisTopic object. It should be a data frame were the rows are cells and the columns data fields.
All cells included in the cisTopic object must be included in this dataframe (additional cells will be ignored).}

\item{...}{Ignored}
}
\value{
Returns a cisTopic object with the new data added in object@cell.data.
}
\description{
Add cells metadata to a given cisTopic object
}
\examples{
cisTopic_mel <- createcisTopicObject(count.matrix = count.matrix)
cisTopic_mel <- addCellMetadata(cisTopic_mel, cell.data = cell.data)
cisTopic_mel

}
