% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializecisTopic.R
\name{addRegionMetadata}
\alias{addRegionMetadata}
\title{Add regions metadata to a cisTopic object}
\usage{
addRegionMetadata(object, region.data, ...)
}
\arguments{
\item{object}{cisTopic object}

\item{region.data}{Additional metadata to add to the cisTopic object. It should be a data frame were the rows are regions and the columns data fields.
All regions included in the cisTopic object must be included in this dataframe (additional cells will be ignored), named by their position coordinates.}

\item{...}{Ignored}
}
\value{
Returns a cisTopic object with the new data added in object@region.data.
}
\description{
Add regions metadata to a given cisTopic object
}
\examples{
cisTopic_mel <- createcisTopicObject(count.matrix = count.matrix)
cisTopic_mel <- addRegionMetadata(cisTopic_mel, region.data = region.data)
cisTopic_mel

}
