% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{annotateRegions}
\alias{annotateRegions}
\title{annotateRegions}
\usage{
annotateRegions(object, txdb, annoDb, ...)
}
\arguments{
\item{object}{Initialized cisTopic object.}

\item{txdb}{TxDb object matching the genome used for mapping}

\item{annoDb}{Annotation package matchng the organism of origin}

\item{...}{See annotatePeak from ChIPseeker}
}
\value{
Peak annotation is added as new columns to object@region.data
}
\description{
Add peak annotation to the regions (e.g. type of region, closest gene, ...)
}
\details{
See \code{SignaturesHeatmap} for estimating the region types per topic.
}
