% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicTopicOperations.R
\name{binarizecisTopics}
\alias{binarizecisTopics}
\title{Binarize cisTopic}
\usage{
binarizecisTopics(object, method = "GammaFit", thrP = 0.99,
  plot = TRUE, cutoffs = NULL)
}
\arguments{
\item{object}{Initialized cisTopic object, after the object@region.data has been filled with the cisTopic scores.}

\item{method}{Method for the binarization, 'GammaFit' or 'Predefined'. If GammaFit, a gamma distribution will be fitted
to the region scores distributions (a probability threshold must be provided); if threshold, a vector with the top number
of regions to be taken per topic must be provided.}

\item{thrP}{Probability threshold to use as cutoff on the probability distribution when using GammaFit as method.}

\item{plot}{If TRUE, plots of the gamma fit per topic will be produced and a barplot with the number of regions taken
per topic.}

\item{cutoffs}{Vector (or scalar) with the top number of regions per topic to take per topic if Predefined is selected
as method. If scalar, the same number of regions is taken per topic.}
}
\value{
A list with the selected regions per topic stored in object@binarized.cisTopics
}
\description{
cisTopic binarization based on cisTopic scores
}
