% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcisTarget.R
\name{binarizedcisTopicsToCtx}
\alias{binarizedcisTopicsToCtx}
\title{binarizedcisTopicsToCtx}
\usage{
binarizedcisTopicsToCtx(object, genome = "hg19", minOverlap = 0.4, ...)
}
\arguments{
\item{object}{Initialized cisTopic object, after the object@binarized.cisTopics has been filled.}

\item{genome}{Genome to which the data has been mapped. The available genomes are hg19, dm3, dm6 and mm9.}

\item{...}{See \code{findOverlap} from GenomicRanges}
}
\value{
Returns the cisTarget regions that correspond to the binarized topics in object@binarized.regions.to.Rct
}
\description{
Convert binarized cisTopics to the corresponding cisTarget coordinates
}
