% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellData_mel.R
\docType{data}
\name{cellData_mel}
\alias{cellData_mel}
\title{Cell data from simulated single cell epigenomes from 5 bulk H3K27Ac melanoma profiles.}
\format{
A data frame with 100 cells and 2 data fields (cellLine, LineType).
#' \describe{
  \item{cellLine}{Cell line from which the cell comes from:  MM001, MM011, MM034, MM029 and MM047}
  \item{LineType}{Phenotypic line type: melanocyte-like or mesenchymal-like}
}
}
\source{
Verfaille, Imrichová & Kalender-Atak et al. (2015, GSE60666)
}
\usage{
cellData_mel
}
\description{
Cell data from simulated single cell epigenomes from 5 bulk H3K27Ac melanoma profiles.
}
\keyword{datasets}
