% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCells.R
\name{cellTopicHeatmap}
\alias{cellTopicHeatmap}
\title{Cell-cisTopic distribution heatmap}
\usage{
cellTopicHeatmap(object, method = "Zscore", colorBy = NULL,
  colVars = list(), col.low = "dodgerblue", col.mid = "floralwhite",
  col.high = "brown1", ...)
}
\arguments{
\item{object}{Initialized cisTopic object, after the object@selected.model has been filled.}

\item{method}{Select the normalization method to use for plotting: 'Zscore' or 'probability'.}

\item{colorBy}{Select the cell metadata used to colour the plots with. By default, all categorical features are used.}

\item{colVars}{List specifying the colors to use for each label in each colouring level}

\item{col.low}{Color to use for lowest topic enrichment}

\item{col.mid}{Color to use for medium topic enrichment}

\item{col.high}{Color to use for high topic enrichment}

\item{...}{See \code{aheatmap} from NMF}
}
\value{
Heatmap clustering cells based on their cell-cisTopic distributions.
}
\description{
Plot cell states based on dimensionality reduction over cell-cisTopic distributions
}
