% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializecisTopic.R
\name{createcisTopicObject}
\alias{createcisTopicObject}
\title{Initialize and setup the cisTopic object}
\usage{
createcisTopicObject(count.matrix, project.name = "cisTopicProject",
  min.cells = 1, min.regions = 1, is.acc = 0, ...)
}
\arguments{
\item{count.matrix}{Count matrix containing cells as rows and regions as columns. The row names must be the coordinates in position format (e.g. chr1:110-610).}

\item{project.name}{Project name (string).}

\item{min.cells}{Minimal number of cells in which the region has to be accessible. By default, all regions accessible in at least one cell are kept.}

\item{min.regions}{Minimal number of regions that have to be accessible within a cell to be kept. By default, all cells with at least one region accessible are kept.}

\item{is.acc}{Number of counts necessary to consider a region as accessible. When using single cell methylation data, this threshold
represents the beta value from which above regions will be considered methylated.}

\item{...}{Ignored}
}
\value{
Returns a cisTopic object with the counts data stored in object@count.matrix.
object@binary.count.matrix, object@cell.names, object@cell.data, object@regions.ranges, object@regions.data are also initialized.
}
\description{
Initializes the cisTopic object from a counts matrix
}
\details{
For single cell methylation data, the input matrix should contain columns as cells, regions as rows and beta values as values instead of counts.
}
\examples{
cisTopic_mel <- createcisTopicObject(count.matrix = count.matrix)
cisTopic_mel

}
