% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializecisTopic.R
\name{createcisTopicObjectFrom10Xmatrix}
\alias{createcisTopicObjectFrom10Xmatrix}
\title{Initialize and setup the cisTopic object starting from CellRanger ATAC output matrix.}
\usage{
createcisTopicObjectFrom10Xmatrix(data_folder, metrics,
  project.name = "cisTopicProject", min.cells = 1, min.regions = 1,
  is.acc = 1, keepCountsMatrix = TRUE, ...)
}
\arguments{
\item{data_folder}{Matrix folder from CellRanger ATAC (/outs/filtered_peak_bc_matrix/, after untar file).''}

\item{metrics}{List of CellRanger barcodes and their metrics (/outs/singlecell.csv)}

\item{project.name}{Project name (string).}

\item{min.cells}{Minimal number of cells in which the region has to be accessible. By default, all regions accessible in at least one cell are kept.}

\item{min.regions}{Minimal number of regions that have to be accessible within a cell to be kept. By default, all cells with at least one region accessible are kept.}

\item{is.acc}{Number of counts necessary to consider a region as accessible.}

\item{keepCountsMatrix}{Whether to keep the counts matrix or not inside the object. For large matrices, we recommend to set this
to FALSE.}
}
\value{
Returns a cisTopic object with the counts data stored in object@count.matrix.
object@binary.count.matrix, object@cell.names, object@cell.data (including counting statistics), object@regions.ranges, object@regions.data are also initialized.
}
\description{
Initializes the cisTopic object from CellRanger ATAC output files and defined regions.
}
\examples{
data_folder <- '/outs/filtered_peak_bc_matrix'
cisTopicObject <- createcisTopicObjectFrom10Xmatrix(data_folder, metrics)
cisTopicObject
}
