% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializecisTopic.R
\name{createcisTopicObjectFromBAM}
\alias{createcisTopicObjectFromBAM}
\title{Initialize and setup the cisTopic object starting from bam files and defined regions}
\usage{
createcisTopicObjectFromBAM(bamfiles, regions,
  project.name = "cisTopicProject", min.cells = 1, min.regions = 1,
  is.acc = 0, paired = FALSE, ...)
}
\arguments{
\item{bamfiles}{List with the paths to the bam files.}

\item{regions}{Path to the bed file with the defined regions.}

\item{project.name}{Project name (string).}

\item{min.cells}{Minimal number of cells in which the region has to be accessible. By default, all regions accessible in at least one cell are kept.}

\item{min.regions}{Minimal number of regions that have to be accessible within a cell to be kept. By default, all cells with at least one region accessible are kept.}

\item{is.acc}{Number of counts necessary to consider a region as accessible.}

\item{paired}{Whether data should be treated as paired end or not. If it is FALSE, we count a read if its 5' end falls within
the region, if false, fragments will be counted instead of individual reads.}

\item{...}{See featureCounts function from Rsubread.}
}
\value{
Returns a cisTopic object with the counts data stored in object@count.matrix.
object@binary.count.matrix, object@cell.names, object@cell.data (including counting statistics), object@regions.ranges, object@regions.data are also initialized.
}
\description{
Initializes the cisTopic object from the raw bam files and defined regions
}
\examples{
bamfiles <- c('example_1.bam', 'example_2.bam', 'example_3.bam')
regions <- 'example.bed'
cisTopicObject <- createcisTopicObjectFromBAM(bamfiles, regions)
cisTopicObject
}
