% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializecisTopic.R
\name{createcisTopicObjectFromMeth}
\alias{createcisTopicObjectFromMeth}
\title{Initialize and setup the cisTopic object starting from methylation files}
\usage{
createcisTopicObjectFromMeth(
  methfiles,
  regions,
  project.name = "cisTopicProject",
  min.cells = 1,
  min.regions = 1,
  is.acc = 0.5,
  ...
)
}
\arguments{
\item{methfiles}{List with the paths to the tab separated files containing the methylation information.}

\item{regions}{Path to the bed file with the defined regions.}

\item{project.name}{Project name (string).}

\item{min.cells}{Minimal number of cells in which the region has to be accessible. By default, all regions accessible in at least one cell are kept.}

\item{min.regions}{Minimal number of regions that have to be accessible within a cell to be kept. By default, all cells with at least one region accessible are kept.}

\item{is.acc}{Minimal beta value to be considered methylated. As default, it is 0.5.}

\item{...}{Ignored}
}
\value{
Returns a cisTopic object with the counts data (in this case, beta values data) stored in object@count.matrix. NA's are filled as 0's.
object@binary.count.matrix, object@cell.names, object@cell.data (including counting statistics), object@regions.ranges, object@regions.data are also initialized.
}
\description{
Initializes the cisTopic object from DNA methylation calls. These files must be tab separated files containing chromosome, position, number of methylated
reads and total number of reads.
}
\examples{
methfiles <- c('meth_1.txt', 'meth_2.txt', 'meth_3.txt')
regions <- 'example.bed'
cisTopicObject <- createcisTopicObjectFromBAM(methfiles, regions)
cisTopicObject
}
