% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exports.R
\docType{data}
\name{dm3_CtxRegions}
\alias{dm3_CtxRegions}
\title{Data frame containing the mapping from ctx regions to genome coordinates for dm3.}
\format{
Data frame containing the mapping from ctx regions to genome coordinates for dm3.
}
\usage{
dm3_CtxRegions
}
\description{
Data frame containing the mapping from ctx regions to genome coordinates for dm3.
}
\keyword{datasets}
