% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exports.R
\docType{data}
\name{dm6_CtxRegions}
\alias{dm6_CtxRegions}
\title{Data frame containing the mapping from ctx regions to genome coordinates for dm6.}
\format{
Data frame containing the mapping from ctx regions to genome coordinates for dm6.
}
\usage{
dm6_CtxRegions
}
\description{
Data frame containing the mapping from ctx regions to genome coordinates for dm6.
}
\keyword{datasets}
