% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicTopicOperations.R
\name{getBedFiles}
\alias{getBedFiles}
\title{getBedFiles}
\usage{
getBedFiles(object, path)
}
\arguments{
\item{object}{Initialized cisTopic object, after object@binarized.cisTopics has been filled.}

\item{path}{Path to save the coordinates of the cisTopic specific regions as bed files.}

\item{seqlengths}{Annotation from where to retrieve the seqlengths}
}
\description{
Saves topics as bed files
}
