% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicTopicOperations.R
\name{getBigwigFiles}
\alias{getBigwigFiles}
\title{getBigwigFiles}
\usage{
getBigwigFiles(object, path, seqlengths)
}
\arguments{
\item{object}{Initialized cisTopic object, after running getRegionScores}

\item{path}{Path to save bigwig files}
}
\description{
Saves topics as bigwigs files
}
