% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcisTarget.R
\name{getCistromeEnrichment}
\alias{getCistromeEnrichment}
\title{getCistromeEnrichment}
\usage{
getCistromeEnrichment(
  object,
  topic,
  TFname,
  annotation = "TF_highConf",
  aucellRankings,
  nCores = 1,
  aucMaxRank = 0.03 * nrow(aucellRankings),
  plot = TRUE,
  ...
)
}
\arguments{
\item{object}{Initialized cisTopic object, after object@cistromes.regions have been filled (see \code{getCistromes()}).}

\item{topic}{Topic number of the cistrome.}

\item{TFname}{Name of the transcription factor linked to the cistrome.}

\item{annotation}{Annotations to be used ('TF_highConf', 'Both'). By default, only the high confidence annotation is used.}

\item{aucellRankings}{Precomputed aucellRankings using \code{cisTopic_buildRankings()}. These rankings are not stored in the cisTopicObject due to their size.}

\item{nCores}{Number of cores to be used for AUCell}

\item{aucMaxRank}{Threshold to calculate the AUC}

\item{plot}{Whether enrichment plot should be done. If yes, parameters for plotFeatures will not be ignored.}

\item{...}{See \code{plotFeatures()}.}
}
\value{
AUC enrichment values for the signature are stored as a column in object@cell.data. If specified, cells coloured by
their AUC enrichment values will be plotted.
}
\description{
Determine topic-specific cistrome enrichment in the cells. If specified, cistrome enrichment can be plotted.
}
\examples{

cisTopicObject <- getCistromeEnrichment(cisTopicObject, annotation = 'Both', nCores=1)
cisTopicObject

}
