% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcisTarget.R
\name{getCistromes}
\alias{getCistromes}
\title{getCistromes}
\usage{
getCistromes(object, annotation = "Both", nCores = 1, ...)
}
\arguments{
\item{object}{Initialized cisTopic object, after object@binarized.RcisTarget and object@region.data$CtxLabels have been filled.}

\item{annotation}{Annotations to be used ('TF_highConf', 'Both'). By default, only the high confidence annotation is used.}

\item{...}{Ignored}
}
\value{
Motif enrichment table is stored in object@binarized.RcisTarget
}
\description{
Get cistromes formed based on motif enrichment with RcisTarget
}
\examples{

cisTopicObject <- getCistromes(cisTopicObject, annotation = 'Both')
cisTopicObject

}
