% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverlapcisTopic.R
\name{getSignaturesRegions}
\alias{getSignaturesRegions}
\title{getSignaturesRegions}
\usage{
getSignaturesRegions(object, signatures, labels, minOverlap = 0.4, ...)
}
\arguments{
\item{object}{Initialized cisTopic object.}

\item{labels}{A character vector with the names to be given to the signatures (in the same order as the signature paths or named)}

\item{minOverlap}{Minimum overlap between the regions to consider them as overlapping regions}

\item{...}{See findOverlaps from GenomicRanges}

\item{signature}{Path to bed file or data frame containing signature regions (or a vector)}
}
\value{
The regions from the dataset that overlap with the signature are returned as a slot to object@signatures
}
\description{
Get peaks found in region signatures given as a bed files and/or dataframes
}
\examples{
regions <- c('regions_1.bed', 'regions_2.bed', 'regions_3.bed')
cisTopicObject <- getSignaturesRegions(cisTopicObject, regions)
cisTopicObject
}
