% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{GREAT}
\alias{GREAT}
\title{GREAT}
\usage{
GREAT(object, genome = "hg19", fold_enrichment = 2, geneHits = 1,
  sign = 0.05, request_interval = 20, ...)
}
\arguments{
\item{object}{Initialized cisTopic object, after object@binarized.cisTopics has been filled.}

\item{genome}{Genome to which the data was aligned (see rGREAT for options).}

\item{fold_enrichment}{Minimum binomial fold enrichment to keep term.}

\item{geneHits}{Minimum number of genes associated to keep term.}

\item{sign}{Maximum adjusted p-value to keep term.}

\item{request_interval}{Time interval for two requests.}

\item{...}{See \code{submitGreatJob} from rGREAT}
}
\value{
Non-empty GREAT results per topic are return as a list to object@binarized.rGREAT
}
\description{
Run rGREAT in the binarized cisTopics
}
\details{
This function works with regions annotated to hg19, hg18, mm10, mm9 and danRer7.
}
\examples{
object <- GREAT(object, request_interval = 10)

}
