% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exports.R
\docType{data}
\name{hg19_CtxRegions}
\alias{hg19_CtxRegions}
\title{Data frame containing the mapping from ctx regions to genome coordinates for hg19.}
\format{Data frame containing the mapping from ctx regions to genome coordinates for hg19.}
\usage{
hg19_CtxRegions
}
\description{
Data frame containing the mapping from ctx regions to genome coordinates for hg19.
}
\keyword{datasets}
