% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunModels.R
\name{logLikelihoodByIter}
\alias{logLikelihoodByIter}
\title{Plots log likelihood in the models per iteration.}
\usage{
logLikelihoodByIter(object, select = NULL, ...)
}
\arguments{
\item{object}{Initialized cisTopic object, after running runCGSModels.}

\item{select}{Plot log-likelihood per iteration for selected topics (as a vector with the number of topics per model).}

\item{...}{Ignored.}
}
\value{
Plots the log likelihood of the models thorugh the different iterations after burnin.
}
\description{
Plots the log likelihood of the different models through iterations. Log likelihood after burnin should be stable, otherwise more iterations
may be required (together with an increase in the burnin).
}
\examples{
bamfiles <- c('example_1.bam', 'example_2.bam', 'example_3.bam')
regions <- 'example.bed'
cisTopicObject <- CreatecisTopicObjectfromBAM(bamfiles, regions)
cisTopicObject <- runCGSModels(cisTopicObject)
cisTopicObject <- logLikelihoodByIter(cisTopicObject)
}
