% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcisTarget.R
\name{makeBackgroundFeather}
\alias{makeBackgroundFeather}
\title{makeBackgroundFeather}
\usage{
makeBackgroundFeather(pathToFeather, subsetRegions = NULL,
  subsetMotifs = NULL, pathToSave, ...)
}
\arguments{
\item{pathToFeather}{Path to the feather database to use. Note that this database has to match the genome used for mapping.}

\item{subsetRegions}{Subset of regions to be re-ranked (optional).}

\item{subsetMotifs}{Subset of motifs to be used from the database (optional).}

\item{pathToSave}{Path to save background database.}

\item{...}{See RcisTarget}
}
\value{
A file containing the feather database.
}
\description{
Create a background cisTarget ranking out of the genome-wide cisTarget regions based on given regions.
}
\details{
The created database can be used as input for \code{topicsRcisTarget()}; using reduced_database = TRUE. Subsetting by 
regions and motifs is optional, but at least one of them as to be filled.
}
\examples{
pathToFeather <- "hg19-regions-1M-9species.all_regions.mc9nr.feather"
subsetRegion <- cisTopicObject@region.data$CtxLabels
pathToSave <- "hg19-regions-subsetMelanoma.feather"
makeBackgroundFeather(pathToFeather, subsetRegion, pathToSave)
}
