% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exports.R
\docType{data}
\name{mm9_CtxRegions}
\alias{mm9_CtxRegions}
\title{Data frame containing the mapping from ctx regions to genome coordinates for mm9.}
\format{
Data frame containing the mapping from ctx regions to genome coordinates for mm9.
}
\usage{
mm9_CtxRegions
}
\description{
Data frame containing the mapping from ctx regions to genome coordinates for mm9.
}
\keyword{datasets}
