% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{ontologyDotPlot}
\alias{ontologyDotPlot}
\title{ontologyDotPlot}
\usage{
ontologyDotPlot(object, top = 5, topics = "all",
  ontology = "GO Biological Process", var.y = "ID",
  var.col = "Binom_Adjp_BH", var.size = "Binom_Fold_Enrichment",
  order.by = "Binom_Fold_Enrichment", col.low = "brown1",
  col.mid = "floralwhite", col.high = "dodgerblue", min.size = 2,
  max.size = 10)
}
\arguments{
\item{object}{Initialized cisTopic object, after object@binarized.rGREAT has been filled.}

\item{top}{Number of top terms per topic to visualize}

\item{topics}{Topics to plot. By default, all of them; if not a vector with the topic numbers must be provided.}

\item{ontology}{Ontology from which terms are selected to visualize.}

\item{var.y}{Varible (factor) to visualize in the y axis. By default it is ID, but it can be also name.}

\item{var.col}{Variable whose values will be visualized with the color scale}

\item{var.size}{Variable whose values will be visualized with the size scale}

\item{order.by}{Order terms in the y axis (descendent order) based on given variable}

\item{col.low}{Lower color in the color scale}

\item{col.mid}{Middle color in the color scale}

\item{col.high}{Higher color in the color scale}

\item{min.size}{Minimum dot size}

\item{max.size}{Maximum dot size}
}
\value{
Dot plot with the topics, terms and the selected enrichments to visualize
}
\description{
Dot plots for visualizing GREAT results
}
