% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCells.R
\name{plotCellStates}
\alias{plotCellStates}
\title{Plot cell states based on dimensionality reduction over cell-cisTopic distributions}
\usage{
plotCellStates(object, method = "tSNE", colorBy = NULL, dim = 2,
  intervals = 10, topic_contr = "Zscore", topics = "all",
  col.low = "dodgerblue", col.mid = "floralwhite", col.high = "brown1",
  labels = 3, colVars = list(), ...)
}
\arguments{
\item{object}{Initialized cisTopic object, after the object@selected.model has been filled.}

\item{method}{Select the dimensionality reduction method to use for plotting: 'tSNE',  'PCA',  'Biplot', 'DiffusionMap'.}

\item{colorBy}{Select the cell metadata used to colour the plots with. By default, all features are used.}

\item{dim}{Dimensions to use in the plot (2 or 3). Biplot is only doable in 2D.}

\item{intervals}{Intervals to apply on the color palette for coloring continuous variables. By default, it is 10.}

\item{topic_contr}{Color by topic distribution ('Zscore' or 'probability').}

\item{topics}{Vector containing the numbers of the topics for which the plot has to be made if topic_contr is not null.}

\item{col.low}{Color to use for lowest topic enrichment}

\item{col.mid}{Color to use for medium topic enrichment}

\item{col.high}{Color to use for high topic enrichment}

\item{labels}{Labels for the Zscore in the continuous variables plots.}

\item{colsVar}{List specifying the colors to use for each label in each colouring level for cell metadata}
}
\value{
Plots cell states based on the dimensionality reduction method selected, coloured by the given metadata (one plot per feature).
}
\description{
Plot cell states based on dimensionality reduction over cell-cisTopic distributions
}
