% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCells.R
\name{plotFeatures}
\alias{plotFeatures}
\title{Plot features (cells or regions) based on dimensionality reduction over cell-topic or topic-region distributions}
\usage{
plotFeatures(object, target, method = "tSNE", colorBy = NULL, dim = 2,
  intervals = 10, topic_contr = NULL, topics = NULL, col.low = "pink",
  col.mid = "red", col.high = "darkred", labels = 3, colVars = list(),
  plot_ly = FALSE, legend = TRUE, cex.legend = 0.7, factor.min = 0.05,
  factor.max = 0.75, cex.dot = 1, ...)
}
\arguments{
\item{object}{Initialized cisTopic object, after the object@dr has been filled.}

\item{target}{Whether dimensionality reduction should be applied on cells ('cell') or regions (region). Note that for speed and clarity
reasons, dimesionality reduction on regions will only be done using the regions assigned to topics with high confidence 
(see binarizecisTopics()).}

\item{method}{Select the dimensionality reduction method to use for plotting: 'tSNE', 'Umap',  'PCA',  'Biplot', 'DM' (for Diffusion Map).}

\item{colorBy}{Select the cell metadata used to colour the plots with. By default, all features are used.}

\item{dim}{Dimensions to use in the plot (2 or 3). Biplot is only doable in 2D.}

\item{intervals}{Intervals to apply on the color palette for coloring continuous variables. By default, it is 10.}

\item{topic_contr}{Color by topic distribution ('Z-score' or 'Probability').}

\item{topics}{Vector containing the numbers of the topics for which the plot has to be made if topic_contr is not null.}

\item{col.low}{Color to use for lowest topic enrichment}

\item{col.mid}{Color to use for medium topic enrichment}

\item{col.high}{Color to use for high topic enrichment}

\item{labels}{Labels for the Z-score in the continuous variables plots}

\item{plot_ly}{Whether plot_ly should be used for the plots}

\item{legend}{Whether plots should be given with a legend. If FALSE, the legend is given in an independent following plot.}

\item{cex.legend}{Size of the legend}

\item{cex.dot}{Size of the dot}

\item{...}{Ignored.}

\item{colsVar}{List specifying the colors to use for each label in each colouring level for cell metadata}
}
\value{
Plots cell states based on the dimensionality reduction method selected, coloured by the given metadata (one plot per feature).
}
\description{
Plot features (cells or regions) based on dimensionality reduction over cell-topic or topic-region distributions
}
