% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunModels.R
\name{predictiveDistribution}
\alias{predictiveDistribution}
\title{Probability of each region in each cell in the data set.}
\usage{
predictiveDistribution(object, ...)
}
\arguments{
\item{object}{Initialized cisTopic object.}

\item{...}{Ignored.}
}
\value{
Returns a matrix where the rows are the regions, the columns the cells, and the values are the probabilities of seeing a
region in a matrix.
}
\description{
Calculates the probability of each region in each cell.
}
\examples{
bamfiles <- c('example_1.bam', 'example_2.bam', 'example_3.bam')
regions <- 'example.bed'
cisTopicObject <- CreatecisTopicObjectfromBAM(bamfiles, regions)
cisTopicObject <- RunModel(cisTopicObject)
cisTopicObject <- selectModel(cisTopicObject)
cisTopicObject <- predictiveDistributioncisTopicObject)
}
