% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializecisTopic.R
\name{renameCells}
\alias{renameCells}
\title{Rename cells in a cisTopic object}
\usage{
renameCells(object, names, ...)
}
\arguments{
\item{object}{cisTopic object}

\item{names}{Vector with the new names to be given to the cells.}

\item{...}{Ignored}
}
\value{
Returns a cisTopic object with updated cell names in the relevant slots.
}
\description{
Rename cells in a cisTopic object
}
\examples{
cisTopic_mel <- createcisTopicObject(count.matrix = count.matrix)
cisTopic_mel <- renameCells(cisTopic_mel, names)

cisTopic_mel

}
