% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcisTarget.R
\name{scoredRegionsToCtx}
\alias{scoredRegionsToCtx}
\title{scoredRegionsToCtx}
\usage{
scoredRegionsToCtx(object, genome, liftOver = NULL, minOverlap = 0.4,
  ...)
}
\arguments{
\item{object}{Initialized cisTopic object, after the object@binarized.cisTopics has been filled.}

\item{genome}{Genome to which the data has been mapped. The available genomes are hg19, dm3, dm6 and mm9.}

\item{liftOver}{GRangesList object containing the original coordinates (in the same format as
object@region.names) in the data set as slot names and the corresponding mapping regions as a GRanges object in the slot.}

\item{...}{See \code{findOverlap} from GenomicRanges}
}
\value{
Returns the region scores per topic in object@region.data
}
\description{
Map regions to the most overlapping cisTarget region
}
