% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverlapcisTopic.R
\name{signatureCellEnrichment}
\alias{signatureCellEnrichment}
\title{signatureCellEnrichment}
\usage{
signatureCellEnrichment(
  object,
  aucellRankings,
  selected.signatures = "all",
  nCores = 1,
  aucMaxRank = 0.03 * nrow(aucellRankings),
  plot = TRUE,
  ...
)
}
\arguments{
\item{object}{Initialized cisTopic object, after object@cistromes.regions have been filled (see \code{getCistromes()}).}

\item{aucellRankings}{Precomputed aucellRankings using \code{AUCell_buildRankings()}. These rankings are not stored in the cisTopicObject due to their size.}

\item{selected.signatures}{By default all signatures will be used, but signatures can be selected based on index or name.
Alternatively, 'annotation' can be selected to use as signatures the region type labels (e.g. promoter, distal intergenic, ...)
For this, the function annotateRegions() must be run first.}

\item{nCores}{Number of cores to be used for AUCell}

\item{aucMaxRank}{Threshold to calculate the AUC}

\item{plot}{Whether enrichment plot should be done. If yes, parameters for plotFeatures will not be ignored.}

\item{...}{See \code{plotFeatures()}.}
}
\value{
AUC enrichment values for the signature are stored as a column in object@cell.data. If specified, cells coloured by
their AUC enrichment values will be plotted.
}
\description{
Determine signatures enrichment in the cells. If specified, signature enrichment can be plotted.
}
\examples{

pred.matrix <- predictiveDistribution(cisTopicObject)
aucellRankings <- AUCell_buildRankings(pred.matrix)
cisTopicObject <-  signatureCellEnrichment(cisTopicObject, aucellRankings, nCores=1)
cisTopicObject

}
