% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverlapcisTopic.R
\name{signaturesHeatmap}
\alias{signaturesHeatmap}
\title{signaturesHeatmap}
\usage{
signaturesHeatmap(object, topics = "all", selected.signatures = "all",
  nCores = 4, aucMaxRank = 0.03 * nrow(aucellRankings),
  col.low = "dodgerblue", col.mid = "floralwhite", col.high = "brown1",
  ...)
}
\arguments{
\item{object}{Initialized cisTopic object, after regions scores have been calculated with getRegionScores and object@signatures
is filled.}

\item{topics}{By default all topics will be used, but topics can be selected based on index.}

\item{selected.signatures}{By default all signatures will be used, but signatures can be selected based on index or name.
Alternatively, 'annotation' can be selected to use as signatures the region type labels (e.g. promoter, distal intergenic, ...)
For this, the function AnnotateRegions must be run first.}

\item{nCores}{Number of cores to be used for AUCell}

\item{aucMaxRank}{Threshold to calculate the AUC}

\item{col.low}{Color to use for lowest signature enrichment}

\item{col.mid}{Color to use for medium signature enrichment}

\item{col.high}{Color to use for high signature enrichment}

\item{...}{See \code{aheatmao} from NMF}
}
\value{
Heatmap showing the enrichment per topic per signature
}
\description{
Heatmap containing the row normalised AUC values for the signatures in the topics.
}
