% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverlapcisTopic.R
\name{topicSignaturesMatrix}
\alias{topicSignaturesMatrix}
\title{topicSignaturesMatrix}
\usage{
topicSignaturesMatrix(
  object,
  topics = "all",
  selected.signatures = "all",
  nCores = 4,
  aucMaxRank = 0.03 * nrow(aucellRankings),
  scale = TRUE,
  ...
)
}
\arguments{
\item{object}{Initialized cisTopic object, after regions scores have been calculated with getRegionScores and object@signatures
is filled.}

\item{topics}{By default all topics will be used, but topics can be selected based on index.}

\item{selected.signatures}{By default all signatures will be used, but signatures can be selected based on index or name.
Alternatively, 'annotation' can be selected to use as signatures the region type labels (e.g. promoter, distal intergenic, ...)
For this, the function annotateRegions() must be run first.}

\item{nCores}{Number of cores to be used for AUCell}

\item{aucMaxRank}{Threshold to calculate the AUC}

\item{scale}{Whether AUC enrichment should be normalized}

\item{...}{See \code{Heatmap} from ComplexHeatmap}
}
\value{
Matrix containing the enrichment per topic per signature
}
\description{
Matrix containing the row normalised AUC values for the signatures in the topics.
}
\examples{
###
# Load cisTopic object:
path2cisTopicObject <- system.file(package="cisTopic", "examples/cisTopicObject_melanoma.Rds")
cisTopicObject <- readRDS(path2cisTopicObject)
###

signatureMatrix <- topicSignaturesMatrix(cisTopicObject)

}
