% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcisTarget.R
\name{binarizedcisTopicsToCtx}
\alias{binarizedcisTopicsToCtx}
\title{binarizedcisTopicsToCtx}
\usage{
binarizedcisTopicsToCtx(object, genome, liftOver = NULL,
  minOverlap = 0.4, ...)
}
\arguments{
\item{object}{Initialized cisTopic object, after the object@binarized.cisTopics has been filled.}

\item{genome}{Genome to which the data has been mapped. If you are using liftOver, provide the genome to which
you data has been lift-overed. The available genomes are hg19, dm3, dm6 and mm9.}

\item{liftOver}{GRangesList object containing the original coordinates (in the same format as
object@region.names) in the data set as slot names and the corresponding mapping regions as a GRanges object in the slot.}

\item{...}{See \code{findOverlap} from GenomicRanges}
}
\value{
Returns the cisTarget regions that correspond to the binarized topics in object@binarized.regions.to.Rct
}
\description{
Convert binarized cisTopics to the corresponding cisTarget coordinates
}
