% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCells.R
\name{cellTopicHeatmap}
\alias{cellTopicHeatmap}
\title{Cell-cisTopic distribution heatmap}
\usage{
cellTopicHeatmap(object, method = "Z-score", colorBy = NULL,
  colVars = NULL, col.low = "floralwhite", col.mid = "pink",
  col.high = "red", select.cells = NULL, ...)
}
\arguments{
\item{object}{Initialized cisTopic object, after the object@selected.model has been filled.}

\item{method}{Select the normalization method to use for plotting: 'Z-score' or 'Probability'.}

\item{colorBy}{Select the cell metadata used to colour the plots with. By default, all categorical features are used.}

\item{colVars}{List specifying the colors to use for each label in each colouring level}

\item{col.low}{Color to use for lowest topic enrichment}

\item{col.mid}{Color to use for medium topic enrichment}

\item{col.high}{Color to use for high topic enrichment}

\item{select.cells}{If a subset of cells want to be used for making the heatmap, selected cell names can be provided (as a vector).}

\item{...}{See \code{Heatmap} from ComplexHeatmap}
}
\value{
Heatmap clustering cells based on their cell-cisTopic distributions.
}
\description{
Plot cell states based on dimensionality reduction over cell-cisTopic distributions
}
\details{
'Z-score' computes the Z-score for each topic assingment per cell/region and 'Probability' divides the topic assignments by the total number
of assignments in the cell/region in the last iteration plus alpha.
}
