% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicTopicOperations.R
\name{getRegionsScores}
\alias{getRegionsScores}
\title{Get region scores}
\usage{
getRegionsScores(object, method = "Zscore", scaled = TRUE)
}
\arguments{
\item{object}{Initialized cisTopic object, after the object@selected.model has been filled.}

\item{method}{Select the normalization method to use for calculating the region score: 'Zscore' or 'probability'.}

\item{rescale}{Whether feature scaling should be performed within the topics. By default, it is set to true.}
}
\value{
Returns the region scores per topic in object@region.data
}
\description{
Get region scores per topic
}
