% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCells.R
\name{runPCA}
\alias{runPCA}
\title{Calculate Principal Components on the cell-cisTopic distributions}
\usage{
runPCA(object, target, method = "Z-score", seed = 123, ...)
}
\arguments{
\item{object}{Initialized cisTopic object, after the object@selected.model has been filled.}

\item{target}{Whether dimensionality reduction should be applied on cells ('cell') or regions (region). Note that for speed and clarity
reasons, dimesionality reduction on regions will only be done using the regions assigned to topics with high confidence 
(see binarizecisTopics()).}

\item{method}{Select the method for processing the cell assignments: 'Z-score' and 'Probability'. In the case of regions, 
an additional method, 'NormTop' is available (see getRegionScores()).}

\item{...}{See \code{prcomp} from the package stats.}
}
\value{
Returns a cisTopic object with a list of PCA information stored in object@dr$cell$PCA or object@dr$region$PCA.
}
\description{
Calculate Principal Components (PCs) on the cell-cisTopic distributions
}
\details{
'Z-score' computes the Z-score for each topic assingment per cell/region. 'Probability' divides the topic assignments by the total number
of assignments in the cell/region in the last iteration plus alpha. If using 'NormTop', regions are given an score defined by: \eqn{\beta_{w, k} (\log
\beta_{w,k} - 1 / K \sum_{k'} \log \beta_{w,k'})}.
}
\section{Slots}{

\describe{
\item{\code{loadings}}{Matrix whose columns contain eigenvectors}

\item{\code{sdev}}{Standard deviations of the PCs}

\item{\code{var.coord}}{Coordinates of the variables (correlation between the variables and the PCs)}

\item{\code{var.cos2}}{Cos2 of the variables. Measures their representation quality.}

\item{\code{var.contrib}}{Contributions of the variables to the PCs}

\item{\code{ind.coord}}{Coordinates of individuals}

\item{\code{ind.cos2}}{Cos2 of the individuals}

\item{\code{ind.contrib}}{Contributions of the individuals to the PCs}

\item{\code{eigs}}{Eigenvalues, which measure the variability retained per PC}

\item{\code{variance.explained}}{Percentage of variance explained by each component}
}}

