% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCells.R
\name{runtSNE}
\alias{runtSNE}
\title{Run tSNE on the cell-cisTopic distributions}
\usage{
runtSNE(object, method = "Zscore", seed = 123, ...)
}
\arguments{
\item{object}{Initialized cisTopic object, after the object@selected.model has been filled.}

\item{method}{Select the method for processing the cell assignments: 'Zscore',  'probability', 'predictive.distribution'.}

\item{seed}{Integer for making the results reproducible.}

\item{...}{See Rtsne from the package Rtsne.}
}
\value{
Returns a cisTopic object with the tSNE coordinates stored in object@dr$tSNE.
}
\description{
Run tSNE on the cell-cisTopic distributions.
}
\details{
Zscore computes the Z-score for each topic assingment per cell, probability divides the topic assignments by the total number
of assignments in the cell in the last iteration plus alpha and predictive distribution calculates the probability of the seeing
the regions per cell.
}
