% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcisTarget.R
\name{scoredRegionsToCtx}
\alias{scoredRegionsToCtx}
\title{scoredRegionsToCtx}
\usage{
scoredRegionsToCtx(object, genome = "hg19", minOverlap = 0.4, ...)
}
\arguments{
\item{object}{Initialized cisTopic object, after the object@binarized.cisTopics has been filled.}

\item{genome}{to which the data has been mapped. The available genomes are hg19, dm3, dm6 and mm9.}

\item{...}{See \code{findOverlap} from GenomicRanges}
}
\value{
Returns the region scores per topic in object@region.data
}
\description{
Map regions to the most overlapping cisTarget region
}
