% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcisTarget.R
\name{topicsRcisTarget}
\alias{topicsRcisTarget}
\title{topicsRcisTarget}
\usage{
topicsRcisTarget(object, genome = "hg19", pathToFeather,
  reduced_database = FALSE, nesThreshold = 3, rocthr = 0.005,
  maxRank = 20000, nCores = 1, ...)
}
\arguments{
\item{object}{Initialized cisTopic object, after the object@binarized.regions.to.Rct has been filled.}

\item{genome}{Genome to which the data was aligned (hg19 or mm9)}

\item{...}{See RcisTarget}

\item{motifRankings}{Feather database corresponding to the genome}
}
\value{
Motif enrichment table is stored in object@binarized.RcisTarget
}
\description{
Run RcisTarget in the binarized topics
}
\examples{
path <- "/media/seq-srv-06/lcb/cbravo/feather_databases/hg19-regions-1M-9species.all_regions.mc9nr.feather"
motifRankings <- importRankings(path)
cisTopicObject <- topicRcisTarget(cisTopicObject, featherdatabase)
cisTopicObject
}
