% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcisTarget.R
\name{topicsRcisTarget}
\alias{topicsRcisTarget}
\title{topicsRcisTarget}
\usage{
topicsRcisTarget(object, org = "hg19", pathToFeather,
  reduced_database = FALSE, nesThreshold = 3, rocthr = 0.005,
  maxRank = 20000, nCores = 1, ...)
}
\arguments{
\item{object}{Initialized cisTopic object, after the object@binarized.regions.to.Rct has been filled.}

\item{org}{Genome to which the data was aligned (hs, mm or dmel).}

\item{pathToFeather}{Path to the feather database to use. Note that this database has to match the genome used for mapping.}

\item{reduced_database}{Whether the reduced version of the database is being used or not (by default, it is set to false).}

\item{...}{See RcisTarget}

\item{motifRankings}{Feather database corresponding to the genome}
}
\value{
Motif enrichment table is stored in object@binarized.RcisTarget
}
\description{
Run RcisTarget in the binarized topics
}
\examples{
path <- "hg19-regions-1M-9species.all_regions.mc9nr.feather"
motifRankings <- importRankings(path)
cisTopicObject <- topicRcisTarget(cisTopicObject, featherdatabase)
cisTopicObject
}
